Fix csv_means script to use pandas and be efficient
The original method from Adam involved looping and appending to lists
that was here before was very slow when operating on Affdex CSV files
that were 20000+ lines of data. The pandas library is made for this kind
of data analysis. So basically everything has been re-written to use
pandas, and it's also more affdex-specific, since a few columns are
removed before processing occurs.